* Warnings in Sight core was mainly due to the call of macro with fewer parameters than required.
* We have now a macro `fwCoreClassMacro` with three versions:
  * fwCoreClassMacro(_class);
  * fwCoreClassMacro(_class, _parentClass);
  * fwCoreClassMacro(_class, _parentClass, _factory);
* The macro `fwCoreServiceClassDefinitionsMacro` used in services has been simplified and replaced by `fwCoreServiceMacro(_class, _parentClass);`
* All old deprecated macros have been placed in the `macros-legacy.hpp` file (still included by `macros.hpp` for now - until **Sight 22.0**)
* List of new projects that no longer generate warnings during compilation (new cmake option `WARNINGS_AS_ERRORS` enabled):
  * arData
  * arDataTest
  * fwCom
  * fwComTest
  * fwCore
  * fwCoreTest
  * fwData
  * fwDataTest
  * fwMedData
  * fwMedDataTest
  * fwRuntime
  * fwRuntimeTest
  * fwServices
  * fwServicesTest
  * fwTools
  * fwToolsTest
  * fwRenderOgre
  * fwRenderOgreTest

Closes #371

See merge request Sight/sight!255

+++ SrcLib/visu/fwRenderOgre/test/tu/src/ShadingTest.cpp
@@ -1,7 +1,7 @@
 /************************************************************************
  *
- * Copyright (C) 2014-2017 IRCAD France
- * Copyright (C) 2014-2017 IHU Strasbourg
+ * Copyright (C) 2014-2019 IRCAD France
+ * Copyright (C) 2014-2019 IHU Strasbourg
  *
  * This file is part of Sight.
  *
@@ -345,7 +345,7 @@ void ShadingTest::createObjectFromShaderParameter()
         value.i = {{ 321, 0, 0, 0 }};
         obj     = Shading::createObjectFromShaderParameter(::Ogre::GpuConstantType::GCT_INT1, value);
         CPPUNIT_ASSERT_EQUAL(std::string("::fwData::Integer"), obj->getClassname());
-        CPPUNIT_ASSERT_EQUAL(321, ::fwData::Integer::dynamicCast(obj)->getValue());
+        CPPUNIT_ASSERT_EQUAL(static_cast<std::int64_t>(321), ::fwData::Integer::dynamicCast(obj)->getValue());
     }
 
     // Array types
